requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2020.1
requests==2.24.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
